+++ lib/string.h
@@ -1,10 +1,12 @@
 #pragma once
 
-int   memcmp(const void*, const void*, int);
-void *memmove(void*, const void*, int);
-void  memcpy(void*, const void*, int);
-void *memset(void*, int, int);
-int   strlen(const char*);
-int   strncmp(const char*, const char*, int);
-char *strncpy(char*, const char*, int);
-char *strrev(char *s);
+#include <stddef.h>
+
+int    memcmp(const void*, const void*, size_t);
+void * memmove(void*, const void*, size_t);
+void   memcpy(void*, const void*, size_t);
+void * memset(void*, int, size_t);
+size_t strlen(const char*);
+int    strncmp(const char*, const char*, size_t);
+char * strncpy(char*, const char*, size_t);
+char * strrev(char *s);
